RFCT: refactor out the hooks from SummaryTable
